+++ changelog.md
@@ -23,3 +23,14 @@
 - Fixed mouse cursor stuck issue. ([#0001](https://github.com/oxiKKK/oxware/issues/1))
 - Added support for C++ exceptions when using manual mapping. See [this commit](https://github.com/oxiKKK/oxware/commit/c29da51696fd0c3c034d821f8261cba139566bfa) for more details.
 - Various crash fixes.
+
+# v1.0.2
+
+## Additions and improvements
+
+- Added vanilla-like crosshair changer! ➕
+
+## Bugfixes
+
+- Changed formatting of CColor class from 0-255 to 0-1 to provide compatibility with the way imgui handle colors.
+